Add reusable Django status endpoint  app
This adds a simple Django app, that has defaults for which checks to run,
with only checking the files that don't exist and DBs as a sensible
default. This means instead of adding a whole app to a project, we just
need to update `INSTALLED_APPS`, `urls.py`, and it's requirements.

This is probably missing more sensible defaults. We should probably add
checks for settings related to all the existing checks, and possibly add
a setting that allows you to add an array of checks with your own
custom checkers.