Add test cases mocking edge cases for file checks.
This covers the cases where permission is denied for reading a file or some other unknown error occurs while trying to find it. For the former an integration test might be better. For the latter this is just to increase coverage as there is no other known way to test this.